+++ src/idkmng/lib.rs
@@ -1,4 +1,3 @@
-pub mod args;
 pub mod config;
 pub mod file;
 pub mod funcs;
